+++ src/RadhydroShockCGS/test_radhydro_shock_cgs.cpp
@@ -331,7 +331,7 @@ auto problem_main() -> int
 
 			double err_norm = 0.;
 			double sol_norm = 0.;
-			for (int i = 0; i < xs_exact.size(); ++i) {
+			for (size_t i = 0; i < xs_exact.size(); ++i) {
 				err_norm += std::abs(Trad_interp[i] - Trad_exact[i]);
 				sol_norm += std::abs(Trad_exact[i]);
 			}
@@ -349,10 +349,10 @@ auto problem_main() -> int
 #ifdef HAVE_PYTHON
 		std::vector<double> xs_scaled(xs.size());
 		std::vector<double> xs_exact_scaled(xs_exact.size());
-		for (int i = 0; i < xs.size(); ++i) {
+		for (size_t i = 0; i < xs.size(); ++i) {
 			xs_scaled.at(i) = xs.at(i) / Lx;
 		}
-		for (int i = 0; i < xs_exact.size(); ++i) {
+		for (size_t i = 0; i < xs_exact.size(); ++i) {
 			xs_exact_scaled.at(i) = xs_exact.at(i) / Lx;
 		}
